Data IO API

Read data from NIR video and tracking network

BehaviorDataNIR.nmp_vec — Function

nmp_vec(pos_feature)

Given a 3x2xN array pos_feature of 2D coordinates of the nose, metacorpus and pharynx, this function returns three arrays:

mn: an array of shape 2xN representing the vector from the metacorpus to the nose for each frame.

mp: an array of shape 2xN representing the vector from the metacorpus to the pharynx for each frame.

mp⊥: an array of shape 2xN representing the vector perpendicular to mp for each frame.

Missing docstring.

Missing docstring for read_h5. Check Documenter's build log for details.

Missing docstring.

Missing docstring for read_pos_feature. Check Documenter's build log for details.

Write video

Missing docstring.

Missing docstring for encode_movie. Check Documenter's build log for details.

BehaviorDataNIR.write_behavior_video — Function

write_behavior_video(path_h5, path_vid=nothing; fps=20, encoder_options=nothing, downsample=true)

Writes a video of the behavior data Arguments ––––-

path_h5: path of the HDF5 file

path_vid: path of the video to be generated. if nothing (default), automatically generates it

fps: frame rate of the video (default: 20 to match the acquisition rate)

downsample: if true downsample by factor of 2

encoder_options: in named tuple (e.g. (crf="23", preset="slow"))

vars: variables to display in the video represented as a tuple (varname, value_arr, color). Default nothing which does not display any text

text_pos: position of variables text (default (5,5))

text_size: size of variables text (default 20)

text_font: font of variables text (default Futura)

text_spacing: vertical spacing of variables text (default 1.05)

Notes on the encoder options

preset: possible options are: ultrafast, superfast, veryfast, faster, fast, medium – default preset, slow, slower, veryslow. crf: The range of the CRF scale is 0–51, where 0 is lossless, 23 is the default, and 51 is worst quality possible. A lower value generally leads to higher quality, and a subjectively sane range is 17–28. Consider 17 or 18 to be visually lossless or nearly so; it should look the same or nearly the same as the input but it isn't technically lossless.

BehaviorDataNIR.add_text_to_image — Function

add_text_to_image(img, text_arr, position, colors, fs, font, spacing)

Adds text to an image.

Arguments

img: image to which text is to be added

text_arr: array of strings containing the text to be added

position: position of the text in the image

colors: array of colors for each text string

fs: font size

font: font face

spacing: vertical spacing between text strings

Returns

new_img: image with text added

BehaviorDataNIR.write_mip_video — Function

write_mip_video(param_path, num_timepts, ch, path_vid=nothing; fps=20, encoder_options=nothing)

Writes maximum-intensity projection video of MHD data.

Arguments:

param_path::Dict: Parameter path dictionary containing path_dir_mhd and get_basename keys

num_timepts: Number of timepoints to create video of

ch: Confocal channel to use

path_vid: Path to output video

fps (default 20): frames per second

encoder_options: in named tuple (e.g. (crf="23", preset="slow"))

Merge data from multiple datasets with the same animal

BehaviorDataNIR.merge_nir_data! — Function

merge_nir_data!(combined_data_dict::Dict, data_dict::Dict, data_dict_2::Dict, param::Dict)

Modifies combined_data_dict to add merged NIR data from two datasets data_dict and data_dict_2, together with a parameter file param for the primary dataset.