Locomotion API

Velocity computation

BehaviorDataNIR.get_velocity! — Function

get_velocity!(data_dict::Dict, param::Dict; prefix::String="")

Gets velocity, speed, reversal, and related variables from data_dict and param. Can add a prefix (default empty string) to all confocal variables.

BehaviorDataNIR.speed — Function

speed(Δx::Array{<:AbstractFloat,1}, Δy::Array{<:AbstractFloat,1}, Δt::AbstractFloat)

Computes speed in mm/s

Arguments

Δx: discrete difference of x

Δy: discrete difference of y

Δt: time interval

speed(x, y; lag::Int, fps=FLIR_FPS)

Computes speed using x, y coordinates

lag determines the interval at which to compute the discrete difference

Arguments

x: list of x

y: list of y

lag: number of time points for discerete difference interval

fps: fps

Missing docstring.

Missing docstring for reversal_state. Check Documenter's build log for details.

BehaviorDataNIR.get_reversal_events — Function

get_reversal_events(param, velocity, t_range, max_t)

Finds reversal events.

Arguments:

param: Dictionary containing the following variables:

rev_len_thresh: Number of consecutive reversal time points necessary for a reversal event

rev_v_thresh: Velocity threshold below which the worm is counted as reversing

velocity: Worm velocity

t_range: Time range over which to compute reversal events

max_t: Maximum time point

BehaviorDataNIR.compute_reversal_times — Function

compute_reversal_times(reversals, max_t)

Computes the duration of each reversal event.

Arguments:

reversals: List of all time points where the worm is reversing

max_t: Maximum time point in dataset

Stage data

BehaviorDataNIR.zero_stage — Function

zero_stage(x::Array{<:AbstractFloat,2})

Sets the first time point to start at 0, 0

Arguments

pos_stage: (x,y) location, 2 by T array where T is len(time points)