Worm Posture API

Spline computation

BehaviorDataNIR.segment_worm! — Function

segment_worm!(model, img_raw, img_label; θ=0.75)

Segment the worm given img (size: (968, 732)).

Arguments

model: unet model

img_raw: raw img

img_label: pre-allocated label array <:Int32. Size: (480,360)

Optional arguments

θ=0.75: unet output threshold

BehaviorDataNIR.compute_worm_spline! — Function

compute_worm_spline!(
    param, path_h5, worm_seg_model, worm_thickness, med_axis_dict, pts_order_dict, is_omega_dict,
    x_array, y_array, nir_worm_angle, eccentricity; timepts="all"
)

Compute the worm spline for a given set of parameters. Writes to most of its input parameters.

Arguments:

param: A dictionary containing the parameters for the worm spline computation.

path_h5: The path to the HDF5 file containing the worm data.

worm_seg_model: The worm segmentation model.

worm_thickness: The thickness of the worm.

med_axis_dict: A dictionary containing where to write medial axis data.

pts_order_dict: A dictionary containing where to write the order of spline points.

is_omega_dict: A dictionary containing where to write values of whether the worm is self-intersecting.

x_array: An array containing the x-coordinates of the worm splines. Will be modified.

y_array: An array containing the y-coordinates of the worm splines. Will be modified.

nir_worm_angle: The angle of the worm.

eccentricity: The eccentricity of the worm.

timepts: The timepoints to compute the worm spline for. Defaults to "all".

Returns:

img_label: An array containing the labeled image.

errors: A dictionary containing any errors that occurred during computation.

BehaviorDataNIR.compute_worm_thickness — Function

compute_worm_thickness(param, path_h5, worm_seg_model, med_axis_dict, is_omega_dict)

This function computes the thickness of the worm at each point along its length.

Arguments:

param: A dictionary containing parameters for the function.

path_h5: The path to the HDF5 file containing the worm images.

worm_seg_model: The segmentation model used to segment the worm images.

med_axis_dict: A dictionary containing the medial axis of the worm at each timepoint.

is_omega_dict: A dictionary indicating whether each timepoint corresponds to an omega turn.

Returns:

dists: A vector containing the thickness of the worm at each point along its length.

BehaviorDataNIR.get_segment_end_matrix — Function

get_segment_end_matrix(param, x_array, y_array)

Computes equally-spaced points along the worm spline.

Arguments:

param: Dictionary of parameters that includes the following values:

num_center_pts: Number of points along the spline

segment_len: Length of each segment

x_array: x-locations of spline at each time point

y_array: y-locations of spline at each time point

BehaviorDataNIR.interpolate_splines! — Function

interpolatesplines!(datadict)

Interpolates worm splines to time points where the spline computation crashed.

Curvature computation

BehaviorDataNIR.get_body_angles! — Function

getbodyangles!(data_dict::Dict, param::Dict; prefix::String="")

Adds body angles to data_dict given param. Can add a prefix (default empty string) to all confocal variables.

BehaviorDataNIR.get_angular_velocity! — Function

get_angular_velocity!(data_dict::Dict, param::Dict; prefix::String="")

Gets angular velocity from data_dict and param. Can add a prefix (default empty string) to all confocal variables.

BehaviorDataNIR.get_curvature_variables! — Function

get_curvature_variables!(data_dict::Dict, param::Dict; prefix::String="")

Gets curvature, head angle, and related variables from data_dict and param. Can add a prefix (default empty string) to all confocal variables.

BehaviorDataNIR.get_nose_curling! — Function

get_nose_curling!(data_dict::Dict, param::Dict; prefix::String="")

Gets self intersection variables from data_dict and param. Can add a prefix (default empty string) to all confocal variables.

BehaviorDataNIR.get_worm_body_angle — Function

get_worm_body_angle(x_array, y_array, segment_end_matrix, pts)

Gets the body angle of a worm between the given points in the spline.

Arguments:

x_array: Segmentation array in the x-dimension

y_array: Segmentation array in the y-dimension

segment_end_matrix: Matrix of consistently-spaced segmentation locations across time points

pts: A list of three points in the spline to get the body angle between.

BehaviorDataNIR.get_worm_vector — Function

get_worm_vector(x_array, y_array, segment_end_matrix, seg_range)

Gets the vector the worm is facing, in the lab xy-coordinate system.

Arguments:

x_array: Segmentation array in the x-dimension

y_array: Segmentation array in the y-dimension

segment_end_matrix: Matrix of consistently-spaced segmentation locations across time points

seg_range: Set of segmentation locations to use to compute the centroid (which will be the average of them)

BehaviorDataNIR.get_tot_worm_curvature — Function

get_tot_worm_curvature(body_angle, min_len; directional::Bool=false)

Computes total worm curvature

Arguments:

body_angle: Array of worm body angles

min_len: If there aren't this many angles at a given time point, interpolate that time point instead of computing it

directional::Bool (default false): Use directional curvature.

BehaviorDataNIR.nose_curling — Function

nose_curling(spline_x, spline_y, segment_end; max_i=1)

Gets the smallest ratio between distance between two points in space and distance between them along the worm's curvature. A value of 0 means that the worm is intersecting itself, while a value of 1 means the worm is a straight line.

Arguments:

spline_x: x positions in worm spline

spline_y: y positions in worm spline

segment_end: equally-spaced segments along worm spline

max_i (default 1): Maximum location along the medium axis to try. For nose curling, use 1.