Confocal to NIR timesyncing API

BehaviorDataNIR.get_timing_info! — Function

get_timing_info!(data_dict::Dict, param::Dict, path_h5::String, h5_confocal_time_lag::Integer)

Initializes all timing and syncing variables into data_dict::Dict given param::Dict, path_h5::String and the h5_confocal_time_lag::Integer

BehaviorDataNIR.sync_timing — Function

sync_timing(path_h5, n_rec=1)

Sync NIR and confocal timing. Returns (n_img_nir, daqmx_ai, daqmx_di, img_metadata)

Arguments

path_h5::String: path to h5 file

n_rec::Int: number of recordings in the file

BehaviorDataNIR.sync_stim — Function

sync_stim(stim, timing_stack, timing_nir)

Align stim to confocal and NIR timing. Returns (stim_to_confocal, stim_to_nir)

Arguments

stim::Vector: stim signal

timing_stack: confocal timing

timing_nir: NIR timing

BehaviorDataNIR.nir_vec_to_confocal — Function

nir_vec_to_confocal(vec, confocal_to_nir, confocal_len)

Bins NIR behavioral data to match the confocal time points.

Arguments:

vec: behavioral data vector. Can be 1D or 2D; if 2D time should be on the columns

confocal_to_nir: confocal to NIR time sync

confocal_len: length of confocal dataset

BehaviorDataNIR.nir_to_confocal_t — Function

nir_to_confocal_t(t, nir_to_confocal)

Converts NIR time point t to confocal time point using nir_to_confocal timesync variable.

BehaviorDataNIR.get_timestamps — Function

get_timestamps(path_h5)

Gets NIR timestamps from the NIR data file