Utilities API

Angle utilities

BehaviorDataNIR.recenter_angle — Function

recenter_angle(angle; ref=0)

Recenters angle to be within pi of a reference angle ref (optional, default 0)

BehaviorDataNIR.local_recenter_angle — Function

local_recenter_angle(angles; delta=10)

Recenters angles to be continuous. delta (default 10) is the timespan of reference angles.

BehaviorDataNIR.vec_to_angle — Function

vec_to_angle(vec)

Converts a vector into an angle in the lab xy-coordinate space.

BehaviorDataNIR.make_vec — Function

make_vec(x::Array{<:AbstractFloat,1}, y::Array{<:AbstractFloat,1})

Creates a vector out of x and y position variables.

Other utilities

BehaviorDataNIR.impute_list — Function

impute_list(x::Array{<:AbstractFloat,1})

Imputes missing data (NaN or missing) with interpolation

Arguments

x: 1D data to impute

BehaviorDataNIR.get_lsqerr — Function

get_lsqerr(fit, raw)

Computes least squares error of a fit.

BehaviorDataNIR.savitzky_golay_filter — Function

savitzky_golay_filter(data, lag; is_derivative::Bool=false, has_inflection::Bool=true, is_angle=false)

Filters data using the Savitzky-Golay algorithm, either for smoothing or differentiating purposes.

Arguments:

data: Data to filter.

lag: Interval on each side to use for filtering.

is_derivative::Bool (optional, default false): Whether to differentiate the data (vs just smoothing)

has_inflection::Bool (optional, default true): Whether the data has inflection points within smoothing interval. If set to true, apply higher-order smoothing function.

is_angle::Bool (optional, default false): Whether the data is an angle (ie -pi=pi)